[RDPHOEN-463/RDPHOEN-464] Remove line sorting in function "buffer_init"
Implemented solution for removing the line sorting with the actual
setting of OLD_STUFF-Defines.
Actually, the solution is just tested for Grayscale (as visible result).
As the counter-part firmware in the DSP is missing yet, this solution has
to be tested with the other video formats.
That's why the old version is still part of the code, the #define s are removed later.

Added #define PIXEL_PER_CYCLE for magic number 16 and added information
documentation for connection between DMA_WORDSIZE and pixel per cycle.
Added some funtion and code documentation.