[RDPHOEN-463/RDPHOEN-593] Driver spec 1
Implemented the ioctl call in EPC660_Imager
(this commit corresponds to the first commit for EPC660-Imager.cpp).
Set the number of frame buffers to "2".
Together with the printouts in the ISR, this version is running stable (18h).
The printouts in videobuf2-v4l2.c doesn't seem to be determining for stability.
Analyzing raw viewer:
Every 2nd raw picture is mirror inverted and horizontally deferred.